Albums for albums, photos for photos
