+++ photoalbums/api.go
@@ -6,3 +6,8 @@ type AddOrRemoveVisibilityRequestBody struct {
 	AddVisibleTo    []primitive.ObjectID `json:"addVisibleTo"`
 	RemoveVisibleTo []primitive.ObjectID `json:"removeVisibleTo"`
 }
+
+type AddOrRemovePhotosRequestBody struct {
+	AddPhotos    []primitive.ObjectID `json:"addPhotos"`
+	RemovePhotos []primitive.ObjectID `json:"removePhotos"`
+}
