+++ photos/controller.go
@@ -206,44 +206,6 @@ func RequireOwner(photoIdParam string) gin.HandlerFunc {
 	}
 }
 
-// router.DELETE("/albums/:id", RequireAlbumOwner("id"), DeleteAlbum)
-func AddAndRemovePhotosToAlbum(c *gin.Context) {
-	albumIdStr := c.Param("id")
-	albumId, err := primitive.ObjectIDFromHex(albumIdStr)
-	userId := utils.CollectIdFromAuthentication(c)
-
-	if err != nil {
-		c.JSON(400, gin.H{"message": "invalid album ID", "error": err.Error()})
-		return
-	}
-
-	requestBody := AddOrRemovePhotosRequestBody{}
-	if err := c.BindJSON(&requestBody); err != nil {
-		c.JSON(400, gin.H{"message": "invalid request body", "error": err.Error()})
-		return
-	}
-
-	err = db.VerifyVisibilityForAllPhotos(&requestBody.AddPhotos, userId)
-	if err != nil {
-		c.JSON(403, gin.H{"message": "not all photos are accessible", "error": err.Error()})
-		return
-	}
-
-	err = db.VerifyVisibilityForAllPhotos(&requestBody.RemovePhotos, userId)
-	if err != nil {
-		c.JSON(403, gin.H{"message": "not all photos are accessible", "error": err.Error()})
-		return
-	}
-
-	err = db.AddOrRemoveAlbumFromManyPhotos(&albumId, &requestBody.AddPhotos, &requestBody.RemovePhotos)
-	if err != nil {
-		c.JSON(500, gin.H{"message": "error", "error": err.Error()})
-		return
-	}
-
-	c.JSON(200, gin.H{"message": "success"})
-}
-
 func getPagingQueryArgs(c *gin.Context) *utils.IntRange[int] {
 	query := c.Request.URL.Query()
 	var err error
